+++ src/audiounit.rs
@@ -9,6 +9,7 @@ use duplicate::duplicate_item;
 use dyn_clone::DynClone;
 use num_complex::Complex64;
 use rsor::Slice;
+use std::fmt::Write;
 
 /// An audio processor with an object safe interface.
 /// Once constructed, it has a fixed number of inputs and outputs.
@@ -44,7 +45,7 @@ pub trait AudioUnit48: Send + Sync + DynClone {
 
     /// Route constants, latencies and frequency responses at `frequency` Hz
     /// from inputs to outputs. Return output signal.
-    fn route(&self, input: &SignalFrame, frequency: f64) -> SignalFrame;
+    fn route(&mut self, input: &SignalFrame, frequency: f64) -> SignalFrame;
 
     /// Return an ID code for this type of unit.
     fn get_id(&self) -> u64;
@@ -79,7 +80,7 @@ pub trait AudioUnit48: Send + Sync + DynClone {
     /// use num_complex::Complex64;
     /// assert_eq!(pass().response(0, 440.0), Some(Complex64::new(1.0, 0.0)));
     /// ```
-    fn response(&self, output: usize, frequency: f64) -> Option<Complex64> {
+    fn response(&mut self, output: usize, frequency: f64) -> Option<Complex64> {
         assert!(output < self.outputs());
         let mut input = new_signal_frame(self.inputs());
         for i in 0..self.inputs() {
@@ -102,7 +103,7 @@ pub trait AudioUnit48: Send + Sync + DynClone {
     /// let db = tick().response_db(0, 220.0).unwrap();
     /// assert!(db < 1.0e-7 && db > -1.0e-7);
     /// ```
-    fn response_db(&self, output: usize, frequency: f64) -> Option<f64> {
+    fn response_db(&mut self, output: usize, frequency: f64) -> Option<f64> {
         assert!(output < self.outputs());
         self.response(output, frequency).map(|r| amp_db(r.norm()))
     }
@@ -118,7 +119,7 @@ pub trait AudioUnit48: Send + Sync + DynClone {
     /// assert_eq!(tick().latency(), Some(1.0));
     /// assert_eq!((tick() >> tick()).latency(), Some(2.0));
     /// ```
-    fn latency(&self) -> Option<f64> {
+    fn latency(&mut self) -> Option<f64> {
         if self.outputs() == 0 {
             return None;
         }
@@ -228,6 +229,116 @@ pub trait AudioUnit48: Send + Sync + DynClone {
         self.tick(&[x, y], &mut output);
         (output[0], output[1])
     }
+
+    /// Memory footprint of this unit in bytes, without counting buffers and other allocations.
+    fn footprint(&self) -> usize;
+
+    /// Print information about this unit into a string.
+    fn display(&mut self) -> String {
+        let mut string = String::new();
+
+        if self.inputs() > 0 && self.outputs() > 0 && self.response(0, 440.0).is_some() {
+            let scope = [
+                b"------------------------------------------------",
+                b"                                                ",
+                b"------------------------------------------------",
+                b"                                                ",
+                b"------------------------------------------------",
+                b"                                                ",
+                b"------------------------------------------------",
+                b"                                                ",
+                b"------------------------------------------------",
+                b"                                                ",
+                b"------------------------------------------------",
+                b"                                                ",
+                b"------------------------------------------------",
+            ];
+
+            let mut scope: Vec<_> = scope.iter().map(|x| x.to_vec()).collect();
+
+            let f: [f64; 48] = [
+                10.0, 20.0, 30.0, 40.0, 50.0, 60.0, 70.0, 80.0, 90.0, 100.0, 120.0, 140.0, 160.0,
+                180.0, 200.0, 250.0, 300.0, 350.0, 400.0, 450.0, 500.0, 600.0, 700.0, 800.0, 900.0,
+                1000.0, 1200.0, 1400.0, 1600.0, 1800.0, 2000.0, 2500.0, 3000.0, 3500.0, 4000.0,
+                4500.0, 5000.0, 6000.0, 7000.0, 8000.0, 9000.0, 10000.0, 12000.0, 14000.0, 16000.0,
+                18000.0, 20000.0, 22000.0,
+            ];
+
+            let r: Vec<_> = f
+                .iter()
+                .map(|&f| (self.response_db(0, f).unwrap(), f))
+                .collect();
+
+            let epsilon_db = 1.0e-2;
+            let max_r = r.iter().fold((-f64::INFINITY, None), {
+                |acc, &x| {
+                    if abs(acc.0 - x.0) <= epsilon_db {
+                        (max(acc.0, x.0), None)
+                    } else if acc.0 > x.0 {
+                        acc
+                    } else {
+                        (x.0, Some(x.1))
+                    }
+                }
+            });
+            let max_db = ceil(max_r.0 / 10.0) * 10.0;
+
+            for i in 0..f.len() {
+                let row = (max_db - r[i].0) / 5.0;
+                let mut j = ceil(row) as usize;
+                let mut c = if row - floor(row) <= 0.5 { b'*' } else { b'.' };
+                while j < scope.len() {
+                    scope[j][i] = c;
+                    j += 1;
+                    c = b'*';
+                }
+            }
+
+            for (row, ascii_line) in scope.into_iter().enumerate() {
+                let line = String::from_utf8(ascii_line).unwrap();
+                if row & 1 == 0 {
+                    let db = round(max_db - row as f64 * 5.0) as i64;
+                    writeln!(&mut string, "{:3} dB {} {:3} dB", db, line, db).unwrap();
+                } else {
+                    writeln!(&mut string, "       {}", line).unwrap();
+                }
+            }
+
+            writeln!(
+                &mut string,
+                "       |   |    |    |     |    |    |     |    |    |"
+            )
+            .unwrap();
+            writeln!(
+                &mut string,
+                "       10  50   100  200   500  1k   2k    5k   10k  20k Hz\n"
+            )
+            .unwrap();
+
+            write!(&mut string, "Peak Magnitude : {:.2} dB", max_r.0).unwrap();
+
+            match max_r.1 {
+                Some(frequency) => {
+                    writeln!(&mut string, " ({} Hz)", frequency as i64).unwrap();
+                }
+                _ => {
+                    string.push('\n');
+                }
+            }
+        }
+
+        writeln!(&mut string, "Inputs         : {}", self.inputs()).unwrap();
+        writeln!(&mut string, "Outputs        : {}", self.outputs()).unwrap();
+        writeln!(
+            &mut string,
+            "Latency        : {:.1} samples",
+            self.latency().unwrap_or(0.0)
+        )
+        .unwrap();
+        writeln!(&mut string, "Footprint      : {} bytes", self.footprint()).unwrap();
+
+        string
+    }
 }
 
 #[duplicate_item(
@@ -273,8 +384,11 @@ where
     fn ping(&mut self, probe: bool, hash: AttoRand) -> AttoRand {
         self.0.ping(probe, hash)
     }
-    fn route(&self, input: &SignalFrame, frequency: f64) -> SignalFrame {
+    fn route(&mut self, input: &SignalFrame, frequency: f64) -> SignalFrame {
         self.0.route(input, frequency)
+    }
+    fn footprint(&self) -> usize {
+        std::mem::size_of::<X>()
     }
 }
 
@@ -389,7 +503,10 @@ impl AudioUnit48 for BigBlockAdapter48 {
     fn ping(&mut self, probe: bool, hash: AttoRand) -> AttoRand {
         self.source.ping(probe, hash)
     }
-    fn route(&self, input: &SignalFrame, frequency: f64) -> SignalFrame {
+    fn route(&mut self, input: &SignalFrame, frequency: f64) -> SignalFrame {
         self.source.route(input, frequency)
+    }
+    fn footprint(&self) -> usize {
+        self.source.footprint()
     }
 }