+++ src/oscillator.rs
@@ -85,7 +85,7 @@ impl<T: Real> AudioNode for Sine<T> {
         self.reset(None);
     }
 
-    fn route(&self, _input: &SignalFrame, _frequency: f64) -> SignalFrame {
+    fn route(&mut self, _input: &SignalFrame, _frequency: f64) -> SignalFrame {
         let mut output = new_signal_frame(self.outputs());
         output[0] = Signal::Latency(0.0);
         output
@@ -186,7 +186,7 @@ impl<T: Real, N: Size<T>> AudioNode for Dsf<T, N> {
         self.reset(None);
     }
 
-    fn route(&self, _input: &SignalFrame, _frequency: f64) -> SignalFrame {
+    fn route(&mut self, _input: &SignalFrame, _frequency: f64) -> SignalFrame {
         let mut output = new_signal_frame(self.outputs());
         output[0] = Signal::Latency(0.0);
         output
@@ -287,7 +287,7 @@ impl<T: Float> AudioNode for Pluck<T> {
         self.initialized = false;
     }
 
-    fn route(&self, _input: &SignalFrame, _frequency: f64) -> SignalFrame {
+    fn route(&mut self, _input: &SignalFrame, _frequency: f64) -> SignalFrame {
         let mut output = new_signal_frame(self.outputs());
         output[0] = Signal::Latency(0.0);
         output